pyramid: Add panel undervolt interface
Panel voltage can be changed on-the-go via the sysfs file /sys/module/board_pyramid_panel/parameters/panel_uv, where panel_uv is a value between 0 and 700mV to undervolt the panel by (stock panel voltage is 3100mV). After the sysfs control is written to with the desired undervolt value (as a multiple of 25mV), turn the screen off and on to apply the new voltage.

Default panel undervolt is 400mV.

Signed-off-by: Sultanxda <sultanxda@gmail.com>